// src/types/Enums/Descriptions/Conflux/ElementalistsDescription.ts
export enum ElementalistsDescription {
  AENAIN = 'Джинны некогда были созданы великими силами из таинственных магических материй. Энейн получился несовершенным, и столетиями работал над тем, чтобы загладить свои недостатки. Усердие и талант к магии воздуха стали для него путёвкой в Сопряжение.',
  BRISSA = 'По меркам джиннов, Брисса ещё совсем молодая. Не закончив обучения, она ушла воевать в Эрафию во время Войн Возрождения. Там, где не хватало знаний, ей приходилось добирать практическим опытом.',
  CIELE = 'Циэль выросла в АвЛи на берегу океана, и древние воды навсегда остались у неё в душе. Она исходила немало земель в поисках того, что влекло её, и нашла новый дом в Сопряжении.',
  GELARE = 'Джелар с раннего детства отличался мудростью, выдающейся даже для эльфа. На зов Сопряжения он отреагировал со свойственной для себя невозмутимостью.',
  GRINDAN = 'Практически все гномы обожают золото. Некоторые от него просто теряют голову. Но никому другому земля не дарит его с такой щедростью, как Гриндану.',
  INTEUS = 'Интей ещё в детстве научился управлять огнём и стал зарабатывать фокусами в бродячем цирке. Поговаривали, что он сын элементаля Огня. Однажды ночью он почувствовал зов и навсегда покинул свой цирк. Теперь он мастерски владеет магией Огня и служит Сопряжению.',
  LABETHA = 'Лабета родилась в бедной крестьянской семье, её родители погибли во время Войн Возрождения. Девочку забрали дендроиды, и она выросла в АвЛи, где открыла в себе власть над силами земли. Услышав зов Сопряжения, она сразу же ответила на него.',
  LUNA = 'Луна была волшебницей в Энроте, и однажды услышала зов Сопряжения, который привёл её в Эрафию. Здесь ей пришлось забыть магию воды, в которой она специализировалась раньше. Её новым призванием стал огонь.'
}